Actualizo social-twitter a Twitter4J 2.2.2

+++ social/twitter/src/main/java/ar/com/zauber/commons/social/twitter/api/streaming/filter/StreamingFilter.java
@@ -33,7 +33,7 @@ public interface StreamingFilter {
     /**
      * The users id to receive public tweets from. Never returns null.
      */
-    int[] getUserIds();
+    long[] getUserIds();
 
     /**
      * Keywords to track. Never returns null.
